requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
dnspython==2.6.1
ecdsa==0.19.0
email-to==0.1.0
email_validator==2.1.1
emailotp==2.1.7
fastapi==0.110.1
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
lxml==5.2.1
Mako==1.3.3
Markdown==3.6
MarkupSafe==2.1.5
passlib==1.7.4
premailer==3.10.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0